Refactor, bug fixes, standarize integer types across codebase

+++ patterns/fallingSnow.h
@@ -8,8 +8,8 @@ struct FallingSnowData {
 
 FallingSnowData fallingSnowData;
 
-void fallingRainbowSnowEffect(CRGB leds[], int numLEDs) {
-  unsigned long currentTime = millis();
+void fallingRainbowSnowEffect(CRGB leds[], uint16_t numLEDs) {
+  uint32_t currentTime = millis();
   if (currentTime > fallingSnowData.lastChangeTime + fallingSnowData.stormDuration) {
     // Randomly choose the storm's intensity phase
     if (random(0, 2) == 0) {
@@ -27,7 +27,7 @@ void fallingRainbowSnowEffect(CRGB leds[], int numLEDs) {
   }
 
   // Dim all LEDs slightly to give the effect of falling
-  for (int i = 0; i < numLEDs; i++) {
+  for (uint16_t i = 0; i < numLEDs; i++) {
     leds[i].fadeToBlackBy(10);
   }
 
@@ -35,7 +35,7 @@ void fallingRainbowSnowEffect(CRGB leds[], int numLEDs) {
 
   // Randomly add new "snowflakes" and calculate their color based on position
   if (random(0, fallingSnowData.chanceOfNewFlake) == 0) {
-    int flakePos = random(0, numLEDs);
+    uint16_t flakePos = random(0, numLEDs);
     uint8_t hue = (fallingSnowData.currentHue + map(flakePos, 0, numLEDs - 1, 0, 256)) % 256;
 
     if (fallingSnowData.calmPhase) {
