+++ patterns/rainbowWaterfallHueRotate.h
@@ -1,17 +1,17 @@
 // Function to create a moving rainbow effect with varying brightness like a waterfall
-void rainbowWaterfallHueRotate(int waveLength, int speed, int numLEDs) {
-  static int position = 0; // Static variable to hold the current position of the wave
+void rainbowWaterfallHueRotate(uint8_t waveLength, uint8_t speed, uint16_t numLEDs) {
+  static int16_t position = 0; // Static variable to hold the current position of the wave
 
   // Loop through each LED to set its color and brightness
-  for (int i = 0; i < numLEDs; i++) {
+  for (uint16_t i = 0; i < numLEDs; i++) {
     // Calculate the hue for each LED, creating a rainbow that moves along the strip
     uint8_t hue = (i * 256 / numLEDs) + position;
     CRGB color = CHSV(hue, 255, 255); // Create color with full saturation and maximum brightness
 
     // Use a sine function to calculate brightness modulation for a wave effect
     // The sine wave is based on the LED's position and the moving position variable
-    int brightness = (sin8(((i * 256 / waveLength) + position) % 256) - 128) * 2;
-    brightness = max(0, brightness); // Ensure the calculated brightness is not negative
+    int16_t brightness = (sin8(((i * 256 / waveLength) + position) % 256) - 128) * 2;
+    brightness = max(static_cast<int16_t>(0), brightness); // Ensure the calculated brightness is not negative
 
     leds[i] = color; // Set the LED's color
     leds[i].fadeToBlackBy(255 - brightness); // Modulate the LED's brightness using fadeToBlackBy
@@ -26,19 +26,19 @@ void rainbowWaterfallHueRotate(int waveLength, int speed, int numLEDs) {
 }
 
 // Function to create a moving rainbow effect with varying brightness like a waterfall in reverse direction
-void reverseRainbowWaterfallHueRotate(int waveLength, int speed, int numLEDs) {
-  static int position = 0; // Static variable to hold the current position of the wave
+void reverseRainbowWaterfallHueRotate(uint8_t waveLength, uint8_t speed, uint16_t numLEDs) {
+  static int16_t position = 0; // Static variable to hold the current position of the wave
 
   // Loop through each LED to set its color and brightness
-  for (int i = 0; i < numLEDs; i++) {
+  for (uint16_t i = 0; i < numLEDs; i++) {
     // Calculate the hue for each LED, creating a rainbow that moves along the strip
     uint8_t hue = ((numLEDs - i) * 256 / numLEDs) + position;
     CRGB color = CHSV(hue, 255, 255); // Create color with full saturation and maximum brightness
 
     // Use a sine function to calculate brightness modulation for a wave effect
     // The sine wave is based on the LED's position and the moving position variable
-    int brightness = (sin8((((numLEDs - i) * 256 / waveLength) + position) % 256) - 128) * 2;
-    brightness = max(0, brightness); // Ensure the calculated brightness is not negative
+    int16_t brightness = (sin8((((numLEDs - i) * 256 / waveLength) + position) % 256) - 128) * 2;
+    brightness = max(static_cast<int16_t>(0), brightness); // Ensure the calculated brightness is not negative
 
     leds[i] = color; // Set the LED's color
     leds[i].fadeToBlackBy(255 - brightness); // Modulate the LED's brightness using fadeToBlackBy
