+++ patterns/staticFullChainRainbow.h
@@ -1,6 +1,6 @@
-void staticFullChainRainbow(CRGB* leds, short numLEDs) {
+void staticFullChainRainbow(CRGB* leds, uint16_t numLEDs) {
   float hueIncrement = 255.0 / numLEDs; // Spread the rainbow across all LEDs
-  for (int i = 0; i < numLEDs; i++) {
+  for (uint16_t i = 0; i < numLEDs; i++) {
     leds[i] = CHSV(hueIncrement * i, 255, 255); // Full brightness for initial setup
   }
 }
